requirements.txt
flask~=3.1.0
firebase-admin~=6.6.0
pyjwt~=2.10.1
argparse~=1.4.0
cryptography~=44.0.0
jupyterlab-pygments~=0.3.0
pip~=21.2.4
attrs~=24.2.0
wheel~=0.37.0
PySocks~=1.7.1
protobuf~=5.29.3
pillow~=11.1.0
docutils~=0.21.2
filelock~=3.17.0
rsa~=4.9
pyasn1~=0.6.1
cffi~=1.17.1
setuptools~=58.0.4
grpcio~=1.70.0
MarkupSafe~=3.0.2
werkzeug~=3.1.3
click~=8.1.8
jinja2~=3.1.4
blinker~=1.9.0
itsdangerous~=2.2.0
proto-plus~=1.26.0
aiohttp~=3.11.14
pyasn1-modules~=0.4.1
cachetools~=5.5.1
certifi~=2024.8.30
urllib3~=1.26.20
packaging~=24.1
requests~=2.32.3
google-api-python-client~=2.160.0
google-crc32c~=1.6.0
importlib-metadata~=8.5.0
firestore~=0.0.8
grpcio-status~=1.70.0
pyparsing~=3.2.1
httplib2~=0.22.0
idna~=3.10
charset-normalizer~=3.4.0
chardet~=5.2.0
pycparser~=2.22
uritemplate~=4.1.1
cachecontrol~=0.14.2
msgpack~=1.1.0
google-auth-httplib2~=0.2.0